Move `storybook-flat-story-titles` plugin
