+++ web/.storybook/main.ts
@@ -1,12 +1,13 @@
 import type { StorybookConfig } from "@storybook/nextjs-vite";
-import { storyNameFromExport, toId } from "storybook/internal/csf";
 
 import { basename, dirname, resolve } from "path";
 
 import { fileURLToPath } from "url";
-import * as ts from "typescript";
+import {
+  flatStoryTitlesPlugin,
+  flattenStoryIndexTitles,
+} from "./storybook-flat-story-titles";
 
-const STORY_FILE_PATTERN = /\.stories\.[cm]?[jt]sx?$/;
 const STORY_EXTENSIONS = "@(js|jsx|mjs|ts|tsx)";
 const DESIGN_COMPONENT_STORIES = [
   "LangfuseIcon/LangfuseIcon",
@@ -20,53 +21,6 @@ const DESIGN_COMPONENT_STORIES = [
  */
 function getAbsolutePath(value: string) {
   return dirname(fileURLToPath(import.meta.resolve(`${value}/package.json`)));
-}
-
-function addInferredStoryTitle(code: string, fileName: string) {
-  const sourceFile = ts.createSourceFile(
-    fileName,
-    code,
-    ts.ScriptTarget.Latest,
-    true,
-    fileName.endsWith(".tsx") || fileName.endsWith(".jsx")
-      ? ts.ScriptKind.TSX
-      : ts.ScriptKind.TS,
-  );
-  let metaObject: ts.ObjectLiteralExpression | undefined;
-
-  const visit = (node: ts.Node) => {
-    if (metaObject) return;
-
-    if (
-      ts.isCallExpression(node) &&
-      ts.isPropertyAccessExpression(node.expression) &&
-      node.expression.name.text === "meta" &&
-      ts.isObjectLiteralExpression(node.arguments[0])
-    ) {
-      const candidate = node.arguments[0];
-      const hasTitle = candidate.properties.some(
-        (property) =>
-          (ts.isPropertyAssignment(property) ||
-            ts.isShorthandPropertyAssignment(property)) &&
-          (ts.isIdentifier(property.name) ||
-            ts.isStringLiteral(property.name)) &&
-          property.name.text === "title",
-      );
-
-      if (!hasTitle) metaObject = candidate;
-      return;
-    }
-
-    ts.forEachChild(node, visit);
-  };
-
-  visit(sourceFile);
-  if (!metaObject) return null;
-
-  const componentTitle = basename(fileName).replace(STORY_FILE_PATTERN, "");
-  const insertionPoint = metaObject.getStart(sourceFile) + 1;
-
-  return `${code.slice(0, insertionPoint)}\n  title: ${JSON.stringify(componentTitle)},${code.slice(insertionPoint)}`;
 }
 
 const config: StorybookConfig = {
@@ -98,34 +52,7 @@ const config: StorybookConfig = {
       titlePrefix: "Playground",
     },
   ],
-  // Storybook computes index and runtime titles separately. Flatten the index
-  // to the story filename; viteFinal injects the same inferred title at runtime.
-  experimental_indexers: async (indexers) =>
-    indexers.map((indexer) => ({
-      ...indexer,
-      createIndex: async (fileName, options) => {
-        const entries = await indexer.createIndex(fileName, options);
-        if (!STORY_FILE_PATTERN.test(fileName)) return entries;
-
-        return entries.map((entry) => {
-          const componentTitle = entry.title?.split("/").at(-1);
-          if (!componentTitle) return entry;
-
-          const title = options.makeTitle(componentTitle);
-          const storyName = storyNameFromExport(entry.exportName);
-          const derivedId = entry.title
-            ? toId(entry.title, storyName)
-            : undefined;
-
-          return {
-            ...entry,
-            title,
-            __id:
-              entry.__id === derivedId ? toId(title, storyName) : entry.__id,
-          };
-        });
-      },
-    })),
+  experimental_indexers: flattenStoryIndexTitles,
   addons: [
     getAbsolutePath("@storybook/addon-a11y"),
     getAbsolutePath("@storybook/addon-docs"),
@@ -145,19 +72,7 @@ const config: StorybookConfig = {
   // pulled in transitively by the table stories). Pointing at the source makes
   // Storybook resolve named exports exactly like the app does.
   viteFinal: async (viteConfig) => {
-    viteConfig.plugins = [
-      {
-        name: "storybook-inferred-playground-title",
-        enforce: "pre",
-        transform(code, id) {
-          const [fileName] = id.split("?");
-          if (!fileName || !STORY_FILE_PATTERN.test(fileName)) return null;
-
-          return addInferredStoryTitle(code, fileName);
-        },
-      },
-      ...(viteConfig.plugins ?? []),
-    ];
+    viteConfig.plugins = [flatStoryTitlesPlugin, ...(viteConfig.plugins ?? [])];
 
     const sharedSrc = resolve(
       dirname(fileURLToPath(import.meta.url)),
